requirements.txt
customtkinter==5.2.0
matplotlib==3.7.1
Pillow==9.5.0
requests==2.28.2
yfinance==0.2.18
transformers==4.30.2
torch==2.0.1